#test Изменение типа свойства класса con_into_db
int get_key  upd into double get_key{get;set;}

+++ templ_any_class/con_into_db.cs
@@ -12,6 +12,7 @@ namespace test_cons
         /// Доступ к свойству key 
         /// создано на копии проекта
         /// </summary>
-        public int get_key { get; set; }
+        public double get_key { get; set; }
+
     }
 }
